requirements.txt

numpy==2.2.0
matplotlib==3.8.4
pillow==11.3.0
pyyaml==6.0.2
